backend/requirements-minimal.txt
Flask==2.2.3
Flask-Cors==3.0.10
Flask-Session==0.4.0
requests==2.31.0
psycopg2-binary==2.9.7
Werkzeug==2.2.3
python-dotenv==1.0.0
PyJWT==2.8.0
